Add "http-proxy" to portrule
First of all: Thank you for the NSE script!
We have tested it against the following vulnerable HTTP server: https://github.com/christophetd/log4shell-vulnerable-app

However, nmap detects the service as "http-proxy" and not "http" or "https" and therefore the NSE script won't run. We think it might make sense to also add "http-proxy" to the portrule.

PS: https://github.com/christophetd/log4shell-vulnerable-app runs on port 8080, so your script acutally ran. But for cases in which a vulnerable server does not run on port 80, 443 or 8080 or is not detected as "http" or "https", the script would return a false negative.